fix: rename BaseIcon to SchoolIcon
