+++ src/common/components/CustomIcon/SchoolIcon/SchoolIcon.tsx
